chore: update `exports` field and separate responsibilities in core
- Create `merge.ts` and `tsup.config.base.ts` files to better separate
- Update `exports` field in `package.json` adding new entry points for
  `merge`, css` and `tsup.config.base`

- Update imports in packages to use the new `tsup.config.base` entry
  point
