+++ packages/tailwindcss-animations/tsup.config.ts
@@ -1,4 +1,4 @@
 import { defineConfig } from "tsup"
-import { tsupConfig } from "@halvaradop/tailwindcss-core"
+import { tsupConfig } from "@halvaradop/tailwindcss-core/tsup.config.base"
 
 export default defineConfig(tsupConfig)
