+++ packages/tailwindcss-core/package.json
@@ -31,14 +31,20 @@
   },
   "homepage": "https://github.com/halvaradop/tailwindcss-utilities#readme",
   "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
+    "./merge": {
+      "types": "./dist/merge.d.ts",
+      "import": "./dist/merge.js",
+      "require": "./dist/merge.cjs"
     },
-    "./utils": {
+    "./css": {
       "types": "./dist/generate-classes.d.ts",
-      "import": "./dist/generate-classes.js"
+      "import": "./dist/generate-classes.js",
+      "require": "./dist/generate-classes.cjs"
+    },
+    "./tsup.config.base": {
+      "types": "./dist/tsup.config.base.d.ts",
+      "import": "./dist/tsup.config.base.js",
+      "require": "./dist/tsup.config.base.cjs"
     },
     "./tsconfig.base.json": "./tsconfig.base.json"
   },
