+++ packages/tailwindcss-core/src/generate-classes.ts
@@ -3,7 +3,6 @@ import tailwindcss from "tailwindcss"
 import minify from "@csstools/postcss-minify"
 import { PluginsConfig } from "tailwindcss/types/config.js"
 import { FlattenArrayType } from "@halvaradop/ts-utility-types"
-import { isObject, isPrimitive } from "@halvaradop/ts-utility-types/utils"
 
 const TAILWIND_UTILITIES_DIRECTIVE = "@tailwind utilities;"
 
@@ -32,32 +31,4 @@ export const extractClasses = (plugin: FlattenArrayType<PluginsConfig>) => {
         return classes.css
     }
     return generateCSS
-}
-
-/**
- * Create a new object by merging two objects which prioritize the object types
- *
- * @param source first object to merge
- * @param target second object to merge
- * @param priority true if the object type should be prioritized, false otherwise
- * @returns merged two objects
- */
-export const merge = <S extends Record<string, unknown>, T extends Record<string, unknown>>(
-    source: S,
-    target: T,
-    priority: boolean = true
-) => {
-    if (priority && isPrimitive(source) && isObject(target)) return target
-    if (priority && isPrimitive(target) && isObject(source)) return source
-    const merged: Record<string, unknown> = { ...source }
-    for (const key in target) {
-        if (!isObject(target[key])) {
-            if ((priority && !isObject(source[key])) || !priority) {
-                merged[key] = target[key]
-            } 
-        } else {
-            merged[key] = merge(source[key] as S, target[key] as T, priority)
-        }
-    }
-    return merged
 }
