+++ packages/tailwindcss-core/tsup.config.ts
@@ -1,4 +1,4 @@
 import { defineConfig } from "tsup"
-import { tsupConfig } from "./src/index.js"
+import { tsupConfig } from "./src/tsup.config.base.js"
 
 export default defineConfig(tsupConfig)
